agentHarness: keep the kernel out of profiles — surface lexicon overlays live in the shared module
The import-boundary sensor caught turns.ts and runtimeHost.ts importing
./profiles/lexicons from the kernel. Overlays are pure product-vocabulary
data keyed by profile-id string, so they move into
shared/agentHarness/productLexicon.ts (a kernel-allowed prefix) and the
profiles module is deleted; no behavior change.
