+++ packages/athena-webapp/convex/_generated/api.d.ts
@@ -41,7 +41,6 @@ import type * as agentHarness_narrativeTrail from "../agentHarness/narrativeTrai
 import type * as agentHarness_profiles_dailyOperations from "../agentHarness/profiles/dailyOperations.js";
 import type * as agentHarness_profiles_dailyOperationsConformance from "../agentHarness/profiles/dailyOperationsConformance.js";
 import type * as agentHarness_profiles_dailyOperationsUiCoverage from "../agentHarness/profiles/dailyOperationsUiCoverage.js";
-import type * as agentHarness_profiles_lexicons from "../agentHarness/profiles/lexicons.js";
 import type * as agentHarness_profiles_syntheticSecondSurface from "../agentHarness/profiles/syntheticSecondSurface.js";
 import type * as agentHarness_profiles_syntheticSecondSurfaceConformance from "../agentHarness/profiles/syntheticSecondSurfaceConformance.js";
 import type * as agentHarness_profiles_syntheticSecondSurfacePorts from "../agentHarness/profiles/syntheticSecondSurfacePorts.js";
@@ -777,7 +776,6 @@ declare const fullApi: ApiFromModules<{
   "agentHarness/profiles/dailyOperations": typeof agentHarness_profiles_dailyOperations;
   "agentHarness/profiles/dailyOperationsConformance": typeof agentHarness_profiles_dailyOperationsConformance;
   "agentHarness/profiles/dailyOperationsUiCoverage": typeof agentHarness_profiles_dailyOperationsUiCoverage;
-  "agentHarness/profiles/lexicons": typeof agentHarness_profiles_lexicons;
   "agentHarness/profiles/syntheticSecondSurface": typeof agentHarness_profiles_syntheticSecondSurface;
   "agentHarness/profiles/syntheticSecondSurfaceConformance": typeof agentHarness_profiles_syntheticSecondSurfaceConformance;
   "agentHarness/profiles/syntheticSecondSurfacePorts": typeof agentHarness_profiles_syntheticSecondSurfacePorts;
